requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.16.4
annotated-types==0.7.0
anthropic==0.62.0
anyio==4.10.0
APScheduler==3.11.0
arch==7.2.0
attrs==25.3.0
backtesting==0.6.5
bcrypt==4.3.0
beautifulsoup4==4.13.4
bokeh==3.7.3
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
clarabel==0.11.1
click==8.2.1
colorama==0.4.6
contourpy==1.3.3
copulas==0.9.0
coverage==7.10.4
curl_cffi==0.13.0
cvxpy==1.7.1
cycler==0.12.1
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
ecdsa==0.19.1
ecos==2.0.14
email_validator==2.2.0
execnet==2.1.1
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
fbm==0.3.0
FLAML==2.3.5
fonttools==4.59.0
frozendict==2.4.6
frozenlist==1.7.0
greenlet==3.2.4
h11==0.16.0
hmmlearn==0.3.3
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
kiwisolver==1.4.8
langchain==0.3.27
langchain-core==0.3.74
langchain-openai==0.3.29
langchain-text-splitters==0.3.9
langsmith==0.4.13
lightgbm==4.6.0
lxml==6.0.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
mdurl==0.1.2
memory-profiler==0.61.0
multidict==6.6.4
multitasking==0.0.12
narwhals==2.1.2
numpy==1.26.4
openai==1.99.5
opentelemetry-api==1.36.0
orjson==3.11.1
osqp==1.0.4
packaging==25.0
panda==0.3.1
pandas==2.3.1
passlib==1.7.4
patsy==1.0.1
peewee==3.18.2
pillow==11.3.0
platformdirs==4.3.8
plotly==5.24.1
pluggy==1.6.0
propcache==0.3.2
protobuf==5.29.5
psutil==7.0.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyautogen==0.2.20
pycparser==2.22
pydantic==2.11.7
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
pyportfolioopt==1.5.6
pytest==8.4.1
pytest-asyncio==1.1.0
pytest-cov==6.2.1
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-mock==3.14.1
pytest-timeout==2.4.0
pytest-xdist==3.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
pywin32==311
PyYAML==6.0.2
redis==6.4.0
regex==2025.7.34
reportlab==4.4.3
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.1.0
rich-toolkit==0.14.9
rignore==0.6.4
rsa==4.9.1
scikit-learn==1.7.1
scipy==1.16.1
scs==3.2.7.post2
sentry-sdk==2.34.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.42
starlette==0.47.2
statsmodels==0.14.5
tenacity==9.1.2
termcolor==3.1.0
threadpoolctl==3.6.0
tiktoken==0.10.0
tornado==6.5.2
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
ujson==5.10.0
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
xgboost==2.1.4
xyzservices==2025.4.0
yarl==1.20.1
yfinance==0.2.65
zipp==3.23.0
zstandard==0.23.0
